Fix Postgres build error
This commit fixes a build error that occurs when updating apt repositories, as one repository within the image doesn't seem to be available. Ignoring this error allows the RUN directive to continue as expected, as that repository doesn't seem to be relevant for the packages that are being installed.

The `ssh`, `rsync`, and `ca-certificates` packages are now also included, as they are necessary for the upgrade script provided by Odoo to work.