Use Odoo version env var for volume and network names
This change allows volumes and networks of multiple copies of this repository to co-exist, as long as their major Odoo version is unique.

This can be expanded further in the future to also include the Postgres major version in the name, which wasn't necessary to add right now. A different naming scheme for volumes and networks might be needed if that is ever to be implemented in the future.